# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/poems": {
      "get": {
        "description": "One page of published poems, newest first unless asked otherwise. Public — no token required, and drafts are unreachable through it.\n\nEach poem arrives with a teaser rather than its body: the first few lines, plus `truncated` saying whether there is more. The whole text is `GET /poems/{slug}`.\n\n`search` matches title and subtitle, case-insensitively, and every whitespace-separated term has to match one of them. Searching *inside* poems is deliberately not offered here — that is Postgres full-text, and it arrives beside this rather than inside it.\n\n`tag` and `author` take slugs and may each be repeated, which reads as 'any of these' — `?tag=sakralne&tag=balady` is both themes, not their intersection. `featured=true` narrows to what an editor has put on the front page.\n\nEvery parameter is optional and an empty one is read as absent. An unrecognised parameter is a 400 rather than something ignored.",
        "operationId": "listPoems",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Which page to return, counting from 1.",
            "schema": {
              "default": 1,
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "description": "How many rows to return. At most 100.",
            "schema": {
              "default": 20,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Matched against title, subtitle, case-insensitively. Whitespace separates terms and every term has to match one of them.",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Which property to order by.",
            "schema": {
              "default": "publishedAt",
              "type": "string",
              "enum": ["publishedAt", "readCount", "title"]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Which direction to sort in.",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": ["asc", "desc"],
              "example": "asc"
            }
          },
          {
            "name": "featured",
            "required": false,
            "in": "query",
            "schema": {
              "minItems": 1,
              "type": "array",
              "items": {
                "type": "boolean",
                "description": "Whether an editor has put it on the front page.",
                "example": false
              }
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "schema": {
              "minItems": 1,
              "type": "array",
              "items": {
                "type": "string",
                "maxLength": 80,
                "description": "A tag's slug. Repeatable.",
                "example": "sakralne"
              }
            }
          },
          {
            "name": "author",
            "required": false,
            "in": "query",
            "schema": {
              "minItems": 1,
              "type": "array",
              "items": {
                "type": "string",
                "maxLength": 120,
                "description": "An author's slug. Repeatable.",
                "example": "orysia-vechirnia"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of poems.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoemPage"
                }
              }
            }
          },
          "400": {
            "description": "The query does not match the schema."
          }
        },
        "summary": "List published poems",
        "tags": ["poems"]
      },
      "post": {
        "description": "Creates a poem owned by the account making the request. There is no field for an author — a poem belongs to whoever is holding the token.\n\n`status` may be omitted, in which case the poem is a DRAFT and nobody but its author sees it. An author may also send PENDING_REVIEW, which puts it in the moderation queue; PUBLISHED is an editor's to set, and is a 403 otherwise.\n\nThe slug is derived from the title, disambiguated with a numeric suffix if that address is taken, and then fixed for the life of the poem — retitling it later does not move it.\n\n`tags` are slugs of themes that already exist. An unknown one is a 400 naming it: themes are curated, and writing a poem is not how a new one is created.",
        "operationId": "createPoem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePoem"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The poem, as written.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudioPoem"
                }
              }
            }
          },
          "400": {
            "description": "The body does not match the schema, or names a theme that does not exist."
          },
          "401": {
            "description": "No valid access token."
          },
          "403": {
            "description": "Only an editor may publish outright."
          },
          "409": {
            "description": "Two requests raced for the same slug. Nothing was created."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Write a poem",
        "tags": ["poems"]
      }
    },
    "/poems/{slug}": {
      "get": {
        "description": "One published poem by slug, with its full text. A draft, a poem waiting in the queue and a slug nobody has used are all the same 404 — telling a stranger which of the three it was would confirm the poem exists.",
        "operationId": "getPoem",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The poem's slug.",
            "schema": {
              "example": "tin-nad-polem",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The poem.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Poem"
                }
              }
            }
          },
          "404": {
            "description": "No published poem has that slug."
          }
        },
        "summary": "Get a poem",
        "tags": ["poems"]
      }
    },
    "/poems/{id}": {
      "patch": {
        "description": "Changes only the fields present in the body; at least one is required. An author may change their own poems, an editor anybody's.\n\nThe slug is not writable and does not follow the title: a poem's address is settled when it is created, because a link somebody has shared is a promise. Neither are the author, the read count, or the publication date — the last of which is stamped by the server the first time the poem is published, and kept if it is taken down and put back.\n\n`status: PUBLISHED` and `featured` are both editors' decisions and a 403 otherwise. `status: DRAFT` on a published poem takes it down, which is the reversible act that has to happen before it can be deleted.\n\n`tags` replaces the whole set rather than adding to it, so `[]` files the poem under nothing and omitting the field leaves its themes alone.",
        "operationId": "updatePoem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The poem's id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePoem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The poem, as changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudioPoem"
                }
              }
            }
          },
          "400": {
            "description": "The id or the body does not match the schema."
          },
          "401": {
            "description": "No valid access token."
          },
          "403": {
            "description": "The poem is somebody else's, or the change is an editor's to make."
          },
          "404": {
            "description": "No poem has that id."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Change a poem",
        "tags": ["poems"]
      },
      "delete": {
        "description": "Permanent, and only available for a poem that is not currently public. A published poem is refused with a 409: readers may have its address, so taking it down (`status: DRAFT`) is a separate and reversible decision that has to be made first.\n\nDeleting takes the poem's themes and its moderation history with it. Deleting a poem that is already gone is a 404, not a no-op.",
        "operationId": "deletePoem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The poem's id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The poem is gone."
          },
          "400": {
            "description": "The id is not a UUID."
          },
          "401": {
            "description": "No valid access token."
          },
          "403": {
            "description": "The poem is somebody else's."
          },
          "404": {
            "description": "No poem has that id."
          },
          "409": {
            "description": "The poem is published. Take it down first."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a poem",
        "tags": ["poems"]
      }
    },
    "/studio/poems": {
      "get": {
        "description": "One page of the poems belonging to the account making the request, most recently saved first. Drafts, poems waiting in the queue, poems published and poems sent back — every state, which is what `GET /poems` cannot answer because it is pinned to the published ones.\n\nThe author is not a parameter and cannot be one: it is the token. There is no query that reaches another account's drafts, which is why `status` is safe to filter on here and absent from the public feed.\n\n`?status=DRAFT&status=REJECTED` accepts several. `search` matches title and subtitle, case-insensitively, and every whitespace-separated term has to match one of them. `?sort=status` orders by the poem's journey — draft, queued, published, rejected — because that is the order the Postgres enum declares.\n\nRows carry the first few lines rather than the whole poem. Ask for one by id to read it in full.",
        "operationId": "listStudioPoems",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Which page to return, counting from 1.",
            "schema": {
              "default": 1,
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "description": "How many rows to return. At most 100.",
            "schema": {
              "default": 20,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Matched against title, subtitle, case-insensitively. Whitespace separates terms and every term has to match one of them.",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Which property to order by.",
            "schema": {
              "default": "updatedAt",
              "type": "string",
              "enum": ["updatedAt", "createdAt", "publishedAt", "title", "status"]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Which direction to sort in.",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": ["asc", "desc"],
              "example": "asc"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "minItems": 1,
              "type": "array",
              "items": {
                "type": "string",
                "enum": ["DRAFT", "PENDING_REVIEW", "PUBLISHED", "REJECTED"],
                "description": "Where the poem is on its way to being public. Only PUBLISHED poems are returned by the public endpoints.",
                "example": "PUBLISHED"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of my poems.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudioPoemPage"
                }
              }
            }
          },
          "400": {
            "description": "The query does not match the schema."
          },
          "401": {
            "description": "No valid access token."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List my poems",
        "tags": ["poems"]
      }
    },
    "/studio/poems/{id}": {
      "get": {
        "description": "The whole poem, whatever state it is in — the text, its status, when it was submitted and when it was last saved.\n\nAn author may read their own poems; an editor may read anybody's, which is what makes reviewing possible: a decision taken on six lines of teaser is not a review, and the `PATCH /poems/{id}` that fixes a line before approving needs the text it is fixing. Somebody else's poem is a 403 for an author, not a 404 — the caller has signed in and typed an id they got from somewhere, and being told it is not theirs is the only answer they can act on.\n\nBy id rather than slug, like the write routes and unlike `GET /poems/{slug}`: the slug is the poem's public address, and this is the studio's key — it exists before a slug is settled and survives the poem being retitled.",
        "operationId": "getStudioPoem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The poem's id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The poem.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudioPoem"
                }
              }
            }
          },
          "400": {
            "description": "The id is not a UUID."
          },
          "401": {
            "description": "No valid access token."
          },
          "403": {
            "description": "The poem is somebody else's."
          },
          "404": {
            "description": "No poem has that id."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Read one poem, in full",
        "tags": ["poems"]
      }
    },
    "/admin/queue": {
      "get": {
        "description": "Every poem in PENDING_REVIEW, oldest submission first. For editors and above.\n\n**The only list on this API that ascends by default**, because that is what a queue is: read newest-first, the poem that has waited longest is the one nobody ever reaches. It is ordered by when the poem entered the queue and not by when it was last saved, so an editor who fixes a line before approving does not push it to the back of the queue they are working through.\n\nThere is no `status` parameter and cannot be one: the endpoint *is* the status. `author` and `tag` take slugs and may be repeated — `?author=vasyl-stus` is everything one poet has waiting.\n\nRows carry the first few lines. Read a poem in full at `GET /studio/poems/{id}`, which an editor may do for anybody's poem, and decide on it at `POST /poems/{id}/approve` or `/reject`.",
        "operationId": "listPoemQueue",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Which page to return, counting from 1.",
            "schema": {
              "default": 1,
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "description": "How many rows to return. At most 100.",
            "schema": {
              "default": 20,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Matched against title, subtitle, case-insensitively. Whitespace separates terms and every term has to match one of them.",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Which property to order by.",
            "schema": {
              "default": "submittedAt",
              "type": "string",
              "enum": ["submittedAt", "createdAt", "title"]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Which direction to sort in.",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": ["asc", "desc"],
              "example": "asc"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "schema": {
              "minItems": 1,
              "type": "array",
              "items": {
                "type": "string",
                "maxLength": 80,
                "description": "A tag's slug. Repeatable.",
                "example": "sakralne"
              }
            }
          },
          {
            "name": "author",
            "required": false,
            "in": "query",
            "schema": {
              "minItems": 1,
              "type": "array",
              "items": {
                "type": "string",
                "maxLength": 120,
                "description": "An author's slug. Repeatable.",
                "example": "orysia-vechirnia"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of waiting poems.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudioPoemPage"
                }
              }
            }
          },
          "400": {
            "description": "The query does not match the schema."
          },
          "401": {
            "description": "No valid access token."
          },
          "403": {
            "description": "The account's role is not high enough."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List poems waiting for review",
        "tags": ["poems"]
      }
    },
...
  },
  "info": {
    "title": "MoodNight API",
    "description": "Read and write endpoints for the MoodNight poetry site. Response shapes come from the zod schemas in @moodnight/shared, the same ones the API validates against.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "moodnight_refresh": {
        "type": "apiKey",
        "in": "cookie",
        "name": "moodnight_refresh"
      }
    },
    "schemas": {
...
      "PoemPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PoemSummary"
            },
            "description": "The rows on this page."
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "How many rows match the search and filters, across all pages.",
            "example": 137
          },
          "page": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "maximum": 9007199254740991,
            "description": "Which page this is, counting from 1.",
            "example": 1
          },
          "perPage": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "maximum": 9007199254740991,
            "description": "How many rows a full page holds.",
            "example": 20
          },
          "pageCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "How many pages `total` divides into. 0 when nothing matches.",
            "example": 7
          }
        },
        "required": ["items", "total", "page", "perPage", "pageCount"],
        "additionalProperties": false,
        "description": "A page of poems, and how many match in total."
      },
      "StudioPoem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "description": "UUIDv7 — time-ordered, so it sorts by creation."
          },
          "slug": {
            "type": "string",
            "description": "The poem's address.",
            "example": "tin-nad-polem"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "example": "Тінь над полем"
          },
          "subtitle": {
            "nullable": true,
            "description": "The line under the title, if any.",
            "example": "із циклу «Спалені листи»",
            "type": "string",
            "maxLength": 200
          },
          "author": {
            "$ref": "#/components/schemas/PoemAuthor"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PoemTag"
            },
            "description": "Themes, possibly none."
          },
          "readCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "How many times it has been opened.",
            "example": 1247
          },
          "featured": {
            "type": "boolean",
            "description": "Whether an editor has put it on the front page.",
            "example": false
          },
          "status": {
            "type": "string",
            "enum": ["DRAFT", "PENDING_REVIEW", "PUBLISHED", "REJECTED"],
            "description": "Where the poem is on its way to being public. Only PUBLISHED poems are returned by the public endpoints.",
            "example": "PUBLISHED"
          },
          "publishedAt": {
            "nullable": true,
            "description": "When the poem became public, or null if it never has.",
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "submittedAt": {
            "nullable": true,
            "description": "When the poem last entered the moderation queue, or null if it never has. Not cleared when it leaves — it says when the poem last asked to be read, which stays true afterwards.",
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "description": "When the poem was first written."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "description": "When it was last saved."
          },
          "body": {
            "type": "string",
            "description": "The poem, newline-separated."
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "subtitle",
          "author",
          "tags",
          "readCount",
          "featured",
          "status",
          "publishedAt",
          "submittedAt",
          "createdAt",
          "updatedAt",
          "body"
        ],
        "additionalProperties": false,
        "description": "A poem as its author sees it, drafts included."
      },
      "CreatePoem": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "example": "Тінь над полем"
          },
          "subtitle": {
            "nullable": true,
            "description": "The line under the title, if any.",
            "example": "із циклу «Спалені листи»",
            "type": "string",
            "maxLength": 200
          },
          "body": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20000,
            "description": "The poem, newline-separated, as it should be read. Plain text — the line breaks are content, not formatting.",
            "example": "Тінь над полем лягла,\nі вітер затих."
          },
          "tags": {
            "description": "Slugs of existing tags, at most 6. Replaces the whole set.",
            "example": ["melankholiia", "nich"],
            "maxItems": 6,
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 80
            }
          },
          "status": {
            "type": "string",
            "enum": ["DRAFT", "PENDING_REVIEW", "PUBLISHED"],
            "description": "DRAFT keeps the poem private, PENDING_REVIEW submits it to the queue, and PUBLISHED makes it public — the last of which is an editor's to set. Rejecting is not here: it carries a note, and that is the queue's own endpoint.",
            "example": "PENDING_REVIEW"
          }
        },
        "required": ["title", "body"],
        "additionalProperties": false,
        "description": "The fields needed to create a poem."
      },
...
    }
  }
}